requirements.txt
gunicorn
Flask==2.3.3
Flask-Login==0.6.3
Flask-JWT-Extended==4.6.0
Flask-WTF==1.2.1
Flask-Migrate==4.0.7
Flask-Mail==0.9.1
SQLAlchemy==2.0.32
psycopg2-binary==2.9.9
alembic==1.13.2
python-dotenv==1.0.1
email-validator==2.2.0
WTForms==3.1.2
itsdangerous==2.2.0
Jinja2==3.1.4
Werkzeug==2.3.8
gunicorn==21.2.0
pytest==8.3.3
requests==2.32.3
redis==5.0.8
boto3==1.35.21

# Images
Pillow==10.4.0
pillow-heif==0.17.0

